only tokenise diffs that don't already exist

+++ tokenize_diffs.py
@@ -1,4 +1,5 @@
 from multiprocessing import Pool
+import os.path
 
 from tokenize import tokenize,filter_diff_lines
 from load_data import load_data, lines_to_files
@@ -13,14 +14,16 @@ total = training+test+validation
 files = lines_to_files(total)
 
 def tokenize_file(file):
-    f = open(file, "r")
-    content = f.read()
-    f.close()
-    tokens = tokenize(filter_diff_lines(content))
     out_file = file.replace("diffs","diffs_tokenized")
-    f = open(out_file, "w")
-    pickle.dump(tokens,f)
-    f.close()
 
-p = Pool(16)
+    if not os.path.isfile(out_file):
+        f = open(file, "r")
+        content = f.read()
+        f.close()
+        tokens = tokenize(filter_diff_lines(content))
+        f = open(out_file, "w")
+        pickle.dump(tokens,f)
+        f.close()
+
+p = Pool(4)
 p.map(tokenize_file,total)
